Adicionar componente Tutorial e integrar ao painel de recompensas; remover arquivo de visão de usuário vazio
