--- src/views/User/Overview/index.vue
@@ -1,12 +1,0 @@
-<template>
-    <div>
-        
-    </div>
-</template>
-
-
-<script>
-export default {
-    
-}
-</script>
